# Conflicts:
#	app/src/main/java/org/piramalswasthya/sakhi/repositories/BenRepo.kt

+++ app/src/main/java/org/piramalswasthya/sakhi/database/room/BeneficiaryIdsAvail.kt
@@ -28,5 +28,5 @@ data class BeneficiaryIdsAvail(
 
     var benId: Long,
 
-    var benRegId: Int = 0
+    var benRegId: Long = 0
 )
